extract duration indicator to separate function
